requirements.txt
nonebot2[fastapi]>=2.3.1
nonebot-adapter-onebot>=2.4.3
matplotlib>=3.9.1
requests>=2.32.3
pydantic>=2.7.4
psutil>=6.0.0
